Shift subwizard to separate class.
